modified Notepad.py file

+++ Notepad.py
@@ -1,5 +1,6 @@
 import datetime
-date = datetime.datetime.today().strftime("%Y-%m-%d / %H:%M:%S")
+import os
+date = datetime.datetime.now().strftime("%Y-%m-%d / %H:%M:%S")
 
 
 # Добавление заметки
@@ -10,6 +11,7 @@ def input_note():
         new_note.insert(i, input(f"Введите {note_elem[i]}: "))
     new_note.insert(2, date)
     export_to_notepad(new_note)
+    print("Заметка добавлена...")
 
 def export_to_notepad(new_note):
     with open('notepad.csv', 'a+', encoding='utf-8') as data:
@@ -17,9 +19,11 @@ def export_to_notepad(new_note):
 
 
 def import_to_notepad():
-    notepad = dict()
+    notepad = dict()        
     with open('notepad.csv', 'r', encoding='utf-8') as data:
         notes = data.readlines()
+        if len(notes) == 0:
+            print("Нет ни одной записи.\nСоздайте заметку.")
         for i in range(len(notes)):
             notepad[i] = notes[i]
     return notepad
@@ -43,42 +47,61 @@ def edit_note():
     main_import_to_notepad()
     with open('notepad.csv', 'r+', encoding='utf-8') as data:
         notes = data.readlines()
-        id_note = int(input('\nВведите номер нужной записи: '))
-        note = notes[id_note-1].split(sep=";")
-        print('\nЧто Вы хотите изменить?')
-        user_choice = int(input('\n1 - Заголовок\n2 - Текст'))
-        for i in range(len(note_elem)):
-            if i == (user_choice-1):    
-                note[i] = input(f'Введите новый {note_elem[i]}: ')
-        note[2] = date
-        notes[id_note-1] = (';'.join(note)+'\n')
-    replace_notepad('Сохранено...', notes)
+        if len(notes)>0:
+            id_note = int(input('\nВведите номер нужной записи: '))
+            if id_note > len(notes): print("Такой заметки не существует.")
+            else:
+                note = notes[id_note-1].split(sep=";")
+                print('\nЧто Вы хотите изменить?')
+                user_choice = int(input('\n1 - Заголовок\n2 - Текст\n'))
+                for i in range(len(note_elem)):
+                    if i == (user_choice-1):    
+                        note[i] = input(f'Введите новый {note_elem[i]}: ')
+                note[2] = date
+                notes[id_note-1] = (';'.join(note)+'\n')
+                replace_notepad('Сохранено...', notes)
 
 # Удаление заметки
 def delete_note():
     main_import_to_notepad()
     with open('notepad.csv', 'r+', encoding='utf-8') as data:
         notes = data.readlines()
-        id_note = int(input('\nВведите номер нужной заметки: '))
-        del notes[id_note-1]
-    replace_notepad('Удалено...', notes)
+        if len(notes)>0:
+            id_note = int(input('\nВведите номер нужной заметки: '))
+            if id_note > len(notes): print("Такой заметки не существует.")
+            else:
+                del notes[id_note-1]
+                replace_notepad('Удалено...', notes)
 
 # Меню
 def notepad_menu():
-    user_choice = input('\n1 - показать все заметки\n2 - создать новую заметку\n3 - удалить заметку\n4 - изменить заметку\n0 - закрыть программу\n')
+    user_choice = input('\n1 - показать все заметки\n2 - создать новую заметку\n3 - удалить заметку\n4 - изменить заметку\nДругой символ - закрыть программу\n')
     if user_choice == '1':
-        main_import_to_notepad()
-        notepad_menu()
+        if os.path.exists("notepad.csv"):
+            main_import_to_notepad()
+            notepad_menu()
+        else:
+            print("Нет ни одной заметки.\nСоздайте первую заметку")
+            notepad_menu()
+        
     elif user_choice == '2':
         input_note()
         notepad_menu()
     elif user_choice == '3':
-        delete_note()
-        notepad_menu()
+        if os.path.exists("notepad.csv"):
+            delete_note()
+            notepad_menu()
+        else:
+            print("Нет ни одной заметки.\nСоздайте первую заметку")
+            notepad_menu()
     elif user_choice == '4':
-        edit_note()
-        notepad_menu()
-    elif user_choice == '0':
+        if os.path.exists("notepad.csv"):
+            edit_note()
+            notepad_menu()
+        else:
+            print("Нет ни одной заметки.\nСоздайте первую заметку")
+            notepad_menu()
+    else:
         exit
